+++ ingenialink/canopen/network.py
@@ -13,12 +13,10 @@ import can
 import canopen
 import ingenialogger
 from can import CanError
+from can.interfaces.kvaser.canlib import CANLIBOperationError
 
 from ingenialink.canopen.register import CanopenRegister
-from ingenialink.canopen.servo import (
-    CANOPEN_SDO_RESPONSE_TIMEOUT,
-    CanopenServo,
-)
+from ingenialink.canopen.servo import CANOPEN_SDO_RESPONSE_TIMEOUT, CanopenServo
 from ingenialink.enums.register import RegCyclicType
 from ingenialink.exceptions import ILError, ILFirmwareLoadError, ILObjectNotExist
 from ingenialink.network import NET_DEV_EVT, NET_PROT, NET_STATE, Network, SlaveInfo
@@ -132,8 +130,8 @@ CAN_BIT_TIMMING = {
 }
 
 
-class CustomIXXATListener(can.Listener):
-    """Custom listener for IXXAT connection.
+class CustomListener(can.Listener):
+    """Custom listener for IXXAT and KVASER connection.
     It is used to ignore the exceptions that occur when
     the error limit is reached.
     """
@@ -145,7 +143,7 @@ class CustomIXXATListener(can.Listener):
         pass
 
     def on_error(self, exc: Exception) -> None:
-        logger.error(f"An exception occurred with the IXXAT connection. Exception: {exc}")
+        logger.error(f"An exception occurred with the IXXAT or KVASER connection. Exception: {exc}")
 
 
 class NetStatusListener(Thread):
@@ -231,7 +229,7 @@ class CanopenNetwork(Network):
         self.__fw_load_errors_enabled = True
 
         self.__connection_args = {
-            "bustype": self.__device,
+            "interface": self.__device,
             "channel": self.__channel,
             "bitrate": self.__baudrate,
         }
@@ -245,6 +243,11 @@ class CanopenNetwork(Network):
             Containing all the detected node IDs.
 
         """
+        if (self.__device, self.__channel) not in self.get_available_devices():
+            raise ILError(
+                f"The {self.__device.upper()} transceiver is not detected. "
+                "Make sure that it's connected and its drivers are installed."
+            )
         is_connection_created = False
         if self._connection is None:
             is_connection_created = True
@@ -321,7 +324,7 @@ class CanopenNetwork(Network):
             self._teardown_connection()
         return slave_info
 
-    def connect_to_slave(  # type: ignore [override]
+    def connect_to_slave(
         self,
         target: int,
         dictionary: str,
@@ -391,8 +394,8 @@ class CanopenNetwork(Network):
         """
         if self._connection is None:
             self._connection = canopen.Network()
-            if self.__device == CAN_DEVICE.IXXAT.value:
-                self._connection.listeners.append(CustomIXXATListener())
+            if self.__device in [CAN_DEVICE.IXXAT.value, CAN_DEVICE.KVASER.value]:
+                self._connection.listeners.append(CustomListener())
             try:
                 self._connection.connect(**self.__connection_args)
             except CanError as e:
@@ -422,7 +425,7 @@ class CanopenNetwork(Network):
             return
         try:
             self._connection.disconnect()
-        except VCIError as e:
+        except (VCIError, CANLIBOperationError) as e:
             logger.error(f"An exception occurred during the teardown connection. Exception: {e}")
         self._connection = None
         logger.info("Tear down connection.")
@@ -448,8 +451,8 @@ class CanopenNetwork(Network):
                 logger.info("Bus flushed")
         except Exception as e:
             logger.error(f"Could not stop guarding. Exception: {e}")
-        if self.__device == CAN_DEVICE.IXXAT.value:
-            self._connection.listeners.append(CustomIXXATListener())
+        if self.__device in [CAN_DEVICE.IXXAT.value, CAN_DEVICE.KVASER.value]:
+            self._connection.listeners.append(CustomListener())
         try:
             self._connection.connect(**self.__connection_args)
             for servo in self.servos:
@@ -458,7 +461,7 @@ class CanopenNetwork(Network):
         except BaseException as e:
             logger.error(f"Connection failed. Exception: {e}")
 
-    def load_firmware(  # type: ignore [override]
+    def load_firmware(
         self,
         target: int,
         fw_file: str,
@@ -1023,14 +1026,14 @@ class CanopenNetwork(Network):
     def is_listener_started(self) -> bool:
         return self.__listener_net_status is not None
 
-    def start_status_listener(self) -> None:  # type: ignore [override]
+    def start_status_listener(self) -> None:
         """Start monitoring network events (CONNECTION/DISCONNECTION)."""
         if self.__listener_net_status is None:
             listener = NetStatusListener(self)
             listener.start()
             self.__listener_net_status = listener
 
-    def stop_status_listener(self) -> None:  # type: ignore [override]
+    def stop_status_listener(self) -> None:
         """Stops the NetStatusListener from listening to the drive."""
         if self._connection is None:
             return
@@ -1074,3 +1077,16 @@ class CanopenNetwork(Network):
 
     def _set_servo_state(self, node_id: int, state: NET_STATE) -> None:
         self.__servos_state[node_id] = state
+
+    @staticmethod
+    def get_available_devices() -> List[Tuple[str, Union[str, int]]]:
+        """Get the available CAN devices and their channels"""
+        unavailable_devices = [CAN_DEVICE.VIRTUAL]
+        if platform.system() == "Windows":
+            unavailable_devices.append(CAN_DEVICE.SOCKETCAN)
+        return [
+            (available_device["interface"], available_device["channel"])
+            for available_device in can.detect_available_configs(
+                [device.value for device in CAN_DEVICE if device not in unavailable_devices]
+            )
+        ]